requirements.txt
boto3==1.27.0
botocore==1.30.0
Jinja2==3.1.2
moto==4.1.12
python-decouple==3.8
requests==2.31.0
responses==0.23.1
urllib3==1.26.16
